Move test-mode-related parts from aplans/schema.py to separate file
